+++ Reimu/Reimu.vcxproj
@@ -146,7 +146,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="leadinFromFile.cpp" />
+    <ClCompile Include="ReimuleadinFromFile.cpp" />
     <ClCompile Include="ReimuBulletLink.cpp" />
     <ClCompile Include="ReimuBackGround.cpp" />
     <ClCompile Include="ReimuGameEnding.cpp" />
